docker_application/requirement.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
mysqlclient==2.2.7
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
